pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pb-analyzer"
version = "0.1.0"
description = "PowerBuilder AS-IS analyzer"
readme = "docs/PROJECT_STRUCTURE.md"
requires-python = ">=3.11,<3.12"
dependencies = ["PyYAML>=6.0"]

[project.scripts]
pb-analyzer = "pb_analyzer.__main__:main"

[project.optional-dependencies]
dev = [
  "pytest==8.3.5",
  "pytest-cov==6.0.0",
  "ruff==0.9.7",
  "mypy==1.15.0",
  "types-PyYAML==6.0.12.20241230",
]

[tool.pytest.ini_options]
addopts = "-q --maxfail=1"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
